Update 'create_building_unit_cadastre_wizard' wizard
